PPv2 SSR: Limit ray depth intersection to 1 unit
Instead of counting all rays that travel behind a z buffer depth value limit intersections to those within a depth range. This can prevent reflected objects appearing infinitely long but will also shorten reflections on longer object (this may not be as noticeable depending on the type of objects often reflected in your scenes)
Use the value of 1 unit from the _ReflectionDepth define for our range.